ft_printf: new version prepared

+++ 03 ft_printf/libft/ft_bzero.c
@@ -6,15 +6,15 @@
 /*   By: min-kang <minguk.gaang@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/19 15:56:25 by min-kang          #+#    #+#             */
-/*   Updated: 2022/03/01 18:51:37 by min-kang         ###   ########.fr       */
+/*   Updated: 2022/03/05 14:52:07 by min-kang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, int n)
 {
-	size_t	i;
+	int		i;
 	char	*str;
 
 	str = (char *)s;
